refactor/reorganize
