+++ data/src/main/kotlin/com/vultisig/wallet/data/securityscanner/ProviderScannerServiceContract.kt
@@ -2,9 +2,10 @@ package com.vultisig.wallet.data.securityscanner
 
 import com.vultisig.wallet.data.models.Chain
 
-interface SecurityScannerServiceContract {
-    suspend fun scanTransaction(transaction: SecurityScannerTransaction)
+interface ProviderScannerServiceContract {
+    suspend fun scanTransaction(transaction: SecurityScannerTransaction): SecurityScannerResult
     fun getProviderName(): String
     fun supportsChain(chain: Chain): Boolean
     fun getSupportedChains(): List<Chain>
+    fun getSupportedFeatures(): List<SecurityScannerFeaturesType>
 }
