+++ data/src/main/kotlin/com/vultisig/wallet/data/securityscanner/SecurityScannerModule.kt
@@ -1,5 +1,7 @@
 package com.vultisig.wallet.data.securityscanner
 
+import com.vultisig.wallet.data.securityscanner.blockaid.BlockaidRpcClient
+import com.vultisig.wallet.data.securityscanner.blockaid.BlockaidRpcClientContract
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
