+++ data/src/main/kotlin/com/vultisig/wallet/data/securityscanner/blockaid/BlockaidExtensions.kt
@@ -1,0 +1,7 @@
+package com.vultisig.wallet.data.securityscanner.blockaid
+
+import com.vultisig.wallet.data.securityscanner.SecurityScannerResult
+
+fun BlockaidTransactionScanResponse.toSecurityScannerResult(): SecurityScannerResult {
+
+}
